Use sleep in test only, to not get in the way of app execution

+++ config/initializers/custom_initializers.rb
@@ -10,8 +10,16 @@ Rails.application.config.to_prepare do |_config|
     eval(File.read(f)) if File.exist?(f)
   end
 
-  # This can be overridden in the app initializer to wrap the sleep and delete_file in an async job
+  # This block can be overridden in the app initializer to wrap the sleep and delete_file in an async job,
+  # something like this:
+  #
+  # CamaleonDeleteFileJob.set(wait: temporal_time).perform_later(file_key)
+  # # put this in app/jobs/camaleon_delete_file_job.rb:
+  # include CamaleonCms::UploaderHelper
+  # cama_uploader.delete_file(file_key)
   CamaleonCmsUploader.delete_block do |settings, cama_uploader, file_key|
+    next unless Rails.env.test?
+
     sleep(settings[:temporal_time])
     cama_uploader.delete_file(file_key)
   end
